Extracting route plot to new component
